chore: lint files

+++ src/Allowances.sol
@@ -11,13 +11,14 @@ import "./libraries/IntLib.sol";
  * @notice Allow more granular alloance setting, supposed to be used by Account
  */
 contract Allowances {
-  
   using IntLib for int;
   // Variables
 
   /// @dev accountId => owner => asset => subId => delegate => allowance
-  mapping(uint => mapping(address => mapping(IAsset => mapping(uint => mapping(address => uint))))) public positiveSubIdAllowance;
-  mapping(uint => mapping(address => mapping(IAsset => mapping(uint => mapping(address => uint))))) public negativeSubIdAllowance;
+  mapping(uint => mapping(address => mapping(IAsset => mapping(uint => mapping(address => uint)))))
+    public positiveSubIdAllowance;
+  mapping(uint => mapping(address => mapping(IAsset => mapping(uint => mapping(address => uint)))))
+    public negativeSubIdAllowance;
 
   /// @dev accountId => owner => asset => delegate => allowance
   mapping(uint => mapping(address => mapping(IAsset => mapping(address => uint)))) public positiveAssetAllowance;
@@ -29,8 +30,8 @@ contract Allowances {
 
   /**
    * @notice Sets bidirectional allowances for all subIds of an asset.
-   *         During a balance adjustment, if msg.sender not ERC721 approved or owner, 
-   *         asset allowance + subId allowance must be >= amount 
+   *         During a balance adjustment, if msg.sender not ERC721 approved or owner,
+   *         asset allowance + subId allowance must be >= amount
    * @param accountId ID of account
    * @param delegate address to assign allowance to
    * @param allowances positive and negative amounts for each asset
@@ -48,9 +49,9 @@ contract Allowances {
     }
   }
 
-  /** 
-   * @notice Sets bidirectional allowances for a specific subId. 
-   *         During a balance adjustment, the subId allowance is decremented first 
+  /**
+   * @notice Sets bidirectional allowances for a specific subId.
+   *         During a balance adjustment, the subId allowance is decremented first
    * @param accountId ID of account
    * @param delegate address to assign allowance to
    * @param allowances positive and negative amounts for each (asset, subId)
@@ -63,28 +64,33 @@ contract Allowances {
   ) internal {
     uint allowancesLen = allowances.length;
     for (uint i; i < allowancesLen; i++) {
-      positiveSubIdAllowance[accountId][owner][allowances[i].asset][allowances[i].subId][delegate] = allowances[i].positive;
-      negativeSubIdAllowance[accountId][owner][allowances[i].asset][allowances[i].subId][delegate] = allowances[i].negative;
+      positiveSubIdAllowance[accountId][owner][allowances[i].asset][allowances[i].subId][delegate] = allowances[i]
+        .positive;
+      negativeSubIdAllowance[accountId][owner][allowances[i].asset][allowances[i].subId][delegate] = allowances[i]
+        .negative;
     }
   }
-
 
   //////////////
   // Spending //
   //////////////
 
-  /** 
+  /**
    * @notice Consume delegate's allowance to update balance
    * @dev revert if delegate has insufficient allowance
    * @param adjustment amount of balance adjustment for an (asset, subId)
    * @param delegate address of msg.sender initiating change
    */
-  function _spendAllowance( // TODO: rename delegate to caller?
-    AccountStructs.AssetAdjustment memory adjustment, address owner, address delegate
+  function _spendAllowance(
+    // TODO: rename delegate to caller?
+    AccountStructs.AssetAdjustment memory adjustment,
+    address owner,
+    address delegate
   ) internal {
-
     /* Early return if amount == 0 */
-    if (adjustment.amount == 0) { return; }
+    if (adjustment.amount == 0) {
+      return;
+    }
 
     /* determine if positive vs negative allowance is needed */
     if (adjustment.amount > 0) {
@@ -95,7 +101,8 @@ contract Allowances {
         delegate,
         adjustment.amount
       );
-    } else { // adjustment.amount < 0
+    } else {
+      // adjustment.amount < 0
       _spendAbsAllowance(
         adjustment.acc,
         negativeSubIdAllowance[adjustment.acc][owner][adjustment.asset][adjustment.subId],
@@ -119,26 +126,27 @@ contract Allowances {
     mapping(address => uint) storage allowancesForSubId,
     mapping(address => uint) storage allowancesForAsset,
     address delegate,
-    int256 amount
+    int amount
   ) internal {
     uint subIdAllowance = allowancesForSubId[delegate];
     uint assetAllowance = allowancesForAsset[delegate];
 
-    uint256 absAmount = amount.abs();
+    uint absAmount = amount.abs();
     /* subId allowances are decremented before asset allowances */
     if (absAmount <= subIdAllowance) {
       allowancesForSubId[delegate] = subIdAllowance - absAmount;
-    } else if (absAmount <= subIdAllowance + assetAllowance) { 
+    } else if (absAmount <= subIdAllowance + assetAllowance) {
       allowancesForSubId[delegate] = 0;
       allowancesForAsset[delegate] = assetAllowance - (absAmount - subIdAllowance);
     } else {
       revert NotEnoughSubIdOrAssetAllowances(
-        address(this), 
-        msg.sender, 
-        accountId, 
+        address(this),
+        msg.sender,
+        accountId,
         amount,
-        subIdAllowance, 
-        assetAllowance);
+        subIdAllowance,
+        assetAllowance
+      );
     }
   }
